requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
joblib==1.3.2
nltk==3.8.1
numpy==1.26.1
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.3
python-dotenv==1.0.0
regex==2023.10.3
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
uvicorn==0.23.2
